+++ src/app/components/positionTable.tsx
@@ -20,7 +20,7 @@ export function PositionsTable({ positions }: { positions: OrderPosition[] }) {
             <Table.Th w={220}>
               {t(locale as Parameters<typeof t>[0], "descriptionLabel")}
             </Table.Th>
-            <Table.Th w={50} ta="right">
+            <Table.Th w={60} ta="right">
               {t(locale as Parameters<typeof t>[0], "listPrice")}
             </Table.Th>
             <Table.Th w={30} ta="right">
@@ -32,7 +32,7 @@ export function PositionsTable({ positions }: { positions: OrderPosition[] }) {
             <Table.Th w={30} ta="right">
               %
             </Table.Th>
-            <Table.Th w={50} ta="right">
+            <Table.Th w={60} ta="right">
               {t(locale as Parameters<typeof t>[0], "netPrice")}
             </Table.Th>
             <Table.Th w={50}>
@@ -103,14 +103,14 @@ export function PositionsTable({ positions }: { positions: OrderPosition[] }) {
                 />
               </Table.Td>
               <Table.Td className={pos.kostenlos ? "text-(--main)" : ""}>
-                {pos.kostenlos ? "Yes" : ""}
+                {pos.kostenlos ? t(locale, "yes") : ""}
               </Table.Td>
               <Table.Td>{pos.posnr}</Table.Td>
               <Table.Td className="whitespace-normal! max-w-xs">
                 {pos.bemerkung}
               </Table.Td>
               <Table.Td className={pos.aktiv ? "text-(--main)" : ""}>
-                {pos.aktiv ? "Yes" : ""}
+                {pos.aktiv ? t(locale, "yes") : ""}
               </Table.Td>
             </Table.Tr>
           ))}